[refactor] Move data model `FeatureCard` to single file
